monoes_apis/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
lxml
